Rename overlap files to cells
