+++ visualizer/src/Project/service/projectMachine.js
@@ -6,12 +6,12 @@ import { pure } from 'xstate/lib/actions';
 import createApiMachine from './apiMachine';
 import createArraysMachine from './arraysMachine';
 import createCanvasMachine from './canvasMachine';
+import createCellsMachine from './cellsMachine';
 import { EventBus, fromEventBus } from './eventBus';
 import createIDBMachine from './idbMachine';
 import createImageMachine from './imageMachine';
 import createLineageMachine from './lineageMachine';
 import createLoadMachine from './loadMachine';
-import createCellsMachine from './overlapsMachine';
 import createSelectMachine from './selectMachine';
 import createSpotsMachine from './spotsMachine';
 import createToolMachine from './tools/toolMachine';
